pyproject.toml
[tool.poetry]
name = "fun_bot"
version = "0.1.0"
description = "A fun bot"
authors = ["DudeBro249 <appdevdeploy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
taskipy = "^1.10.2"

[tool.taskipy.tasks]
start = "python -m fun_bot"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
